style: move utils to util package
