+++ kotlin/src/main/kotlin/de/frohnmeyerwds/mima/util/U24Array.kt
@@ -1,4 +1,4 @@
-package de.frohnmeyerwds.mima
+package de.frohnmeyerwds.mima.util
 
 class U24Array constructor(val value: IntArray) {
     constructor(size: Int) : this(IntArray(size))
